+++ src/components/dialogs/constants.tsx
@@ -1,4 +1,2 @@
 export const defaultDialogPaddingXClass = 'px-3';
 export const defaultDialogPaddingYClass = 'py-2a';
-export const fullScreenBreakPoint = 576;
-export const defaultWidth = 450;